Update versioning and publishing workflow (#4)
I refined the scripts a little bit in order to make it easier to publish. There were some issues with the previous scripts.

Now, running...

```
npm version
```

Will bump the package.json version, compile the css, js, and html, then stage and commit the changes.

Then, all that would need to be done to publish this new version is...

```
npm publish
```

This will push all tags in the repo to GitHub then publish the latest tag/version to the NPM registry. Happy to accept suggestions to make it easier!